Move keyhandling to parent class, update subclasses correspondingly.
